+++ libs/usd-vector-authoring/include/usdvector/authoring/usd_authoring.h
@@ -2,11 +2,24 @@
 
 #include "usdvector/authoring/authoring.h"
 
+#include <optional>
+
 #include <pxr/pxr.h>
 #include <pxr/usd/usd/stage.h>
 
 namespace usdvector::authoring {
 
-Result<pxr::UsdStageRefPtr> BuildUsdStage(const AuthoringPlan& plan);
+enum class StageUpAxis {
+    Y,
+    Z,
+};
+
+struct StageOptions {
+    std::optional<StageUpAxis> upAxis;
+    std::optional<double> metersPerUnit;
+};
+
+Result<pxr::UsdStageRefPtr> BuildUsdStage(const AuthoringPlan& plan,
+                                          const StageOptions& options = {});
 
 }  // namespace usdvector::authoring